move: SortType Enum 파일 위치 변경
- Product 폴더에서 Auction 폴더로 이동

(#28)
