+++ src/main/java/org/chzz/market/domain/auction/entity/SortType.java
@@ -1,0 +1,8 @@
+package org.chzz.market.domain.auction.entity;
+
+/**
+ * 경매 목록 정렬 타입
+ */
+public enum SortType {
+    POPULARITY, EXPENSIVE, CHEAP, NEWEST;
+}
