+++ src/main/java/org/chzz/market/domain/auction/repository/AuctionRepositoryCustom.java
@@ -2,7 +2,7 @@ package org.chzz.market.domain.auction.repository;
 
 import org.chzz.market.domain.auction.dto.AuctionResponse;
 import org.chzz.market.domain.product.entity.Product.Category;
-import org.chzz.market.domain.product.entity.SortType;
+import org.chzz.market.domain.auction.entity.SortType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
