+++ src/main/java/org/chzz/market/domain/auction/repository/AuctionRepositoryImpl.java
@@ -5,10 +5,10 @@ import static org.chzz.market.domain.auction.entity.QAuction.auction;
 import static org.chzz.market.domain.bid.entity.QBid.bid;
 import static org.chzz.market.domain.image.entity.QImage.image;
 import static org.chzz.market.domain.product.entity.QProduct.product;
-import static org.chzz.market.domain.product.entity.SortType.CHEAP;
-import static org.chzz.market.domain.product.entity.SortType.EXPENSIVE;
-import static org.chzz.market.domain.product.entity.SortType.NEWEST;
-import static org.chzz.market.domain.product.entity.SortType.POPULARITY;
+import static org.chzz.market.domain.auction.entity.SortType.CHEAP;
+import static org.chzz.market.domain.auction.entity.SortType.EXPENSIVE;
+import static org.chzz.market.domain.auction.entity.SortType.NEWEST;
+import static org.chzz.market.domain.auction.entity.SortType.POPULARITY;
 
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
@@ -21,10 +21,9 @@ import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.chzz.market.domain.auction.dto.AuctionResponse;
 import org.chzz.market.domain.auction.dto.QAuctionResponse;
-import org.chzz.market.domain.auction.entity.Auction.Status;
 import org.chzz.market.domain.image.entity.QImage;
 import org.chzz.market.domain.product.entity.Product.Category;
-import org.chzz.market.domain.product.entity.SortType;
+import org.chzz.market.domain.auction.entity.SortType;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.support.PageableExecutionUtils;
