--- src/main/java/org/chzz/market/domain/product/entity/SortType.java
@@ -1,8 +1,0 @@
-package org.chzz.market.domain.product.entity;
-
-/**
- * 상품 정렬 타입
- */
-public enum SortType {
-    POPULARITY, EXPENSIVE, CHEAP, NEWEST;
-}
